refactor(cli): remove upload default
