+++ packages/models/src/lib/implementation/constants.ts
@@ -3,4 +3,3 @@ import { Format } from '../persist-config';
 export const PERSIST_OUTPUT_DIR = '.code-pushup';
 export const PERSIST_FORMAT = ['json' as Format];
 export const PERSIST_FILENAME = 'report';
-export const UPLOAD_SERVER = 'https://example.com/api';
